requirements.txt
Django
djangorestframework
psycopg2
drf-spectacular
channels
daphne
djangorestframework-simplejwt
channels-redis
